I/O feature

+++ Employee.java
@@ -1,8 +1,12 @@
+/**
+ * @author Ritarka samanta
+ * @version 1.0
+ */
 public class Employee {
     private String name;
-    private int salary;
+    private double salary;
 
-    public Employee(String name, int salary) {
+    public Employee(String name, double salary) {
         this.name = name;
         this.salary = salary;
     }
